[JIT] Add registry for backend lowering functions
Summary:
This commit adds a registry for storing lowering functions for backends.
Instead of backends registering these lowering functions in separate C
extension modules, these will be registered in the Torch extension.
Backends are registered statically, so a registry is needed to hold
these lowering functions until Python bindings are created.

Test Plan:
python test/test_jit.py TestBackends
